VMs can now be created locally providing a json file
VMs can be created using either a json file or getting the json from UCSB website
Provides a clear cli to generate VMs
Generic refactoring splitting code in modules
