+++ vmcreator/secrets.py
@@ -1,0 +1,1 @@
+API_SECRETS = ''
